Second import of some freebsd 13-stable features from October 28

+++ sys/dev/aic7xxx/aic79xx_inline.h
@@ -40,7 +40,6 @@
  * POSSIBILITY OF SUCH DAMAGES.
  *
  * $Id: //depot/aic7xxx/aic7xxx/aic79xx_inline.h#57 $
- *
  */
 
 #ifndef _AIC79XX_INLINE_H_
@@ -566,7 +565,7 @@ ahd_inq(struct ahd_softc *ahd, u_int port)
 	return ((ahd_inb(ahd, port))
 	      | (ahd_inb(ahd, port+1) << 8)
 	      | (ahd_inb(ahd, port+2) << 16)
-	      | (ahd_inb(ahd, port+3) << 24)
+	      | (((uint64_t)ahd_inb(ahd, port+3)) << 24)
 	      | (((uint64_t)ahd_inb(ahd, port+4)) << 32)
 	      | (((uint64_t)ahd_inb(ahd, port+5)) << 40)
 	      | (((uint64_t)ahd_inb(ahd, port+6)) << 48)
@@ -968,7 +967,6 @@ ahd_intr(struct ahd_softc *ahd)
 	} else if ((intstat & (PCIINT|SPLTINT)) != 0) {
 		ahd->bus_intr(ahd);
 	} else {
-
 		if ((intstat & SEQINT) != 0)
 			ahd_handle_seqint(ahd, intstat);
